umd: tests: Add normalize flag to nvdla_runtime
Some models require that the data the should be
in range(0,255) where as some require in range(0,1)
based on the data on which the model was trained.

We can now dynamically decide whether to normalize
the data or not by feeding the right normalize value.

For example,
ResNet-50 needs the data to be in range (0, 255)
./nvdla_runtime --loadable <loadableName> --image <imageName> --normalize 1

LeNet needs the data in range(0,1)
./nvdla_runtime --loadable <loadableName> --image <imageName> --normalize 255

By default the normalize value is set at 1

Change-Id: Icee154af32c60fb7fa658c8530e394ec4ffeb28a
Signed-off-by: Sharif Inamdar <isharif@nvidia.com>
Reviewed-by: Prashant Gaikwad <pgaikwad@nvidia.com>